ITS ALIVE!!!!!!!!!!!!!!!!!!!!!
hey there guys, i have got it to post and show the main page now. You ahve to use npm run build then you ahve to go for npm run develop and it should work fine. Now we need to work on how it looks and making sure it gets to each page correctly then we need to make sure it saves the data correctly.. Then we should be good.